+++ src/components/TeacherComponents/DownloadCardGroup/DownloadCardGroup.tsx
@@ -26,6 +26,7 @@ type DownloadCardProps = {
   downloadType: string;
   triggerForm: () => void;
   hasError: boolean;
+  isDownloadsExperiment?: boolean;
 };
 const DownloadCard: FC<DownloadCardProps> = ({
   control,
@@ -33,6 +34,7 @@ const DownloadCard: FC<DownloadCardProps> = ({
   downloadType,
   triggerForm,
   hasError,
+  isDownloadsExperiment,
 }) => {
   return (
     <Controller
@@ -73,7 +75,7 @@ const DownloadCard: FC<DownloadCardProps> = ({
             onBlur={onBlur}
             hasError={hasError}
             data-testid={`download-card-${downloadType}`}
-            useDownloadsExperiment={true}
+            isDownloadsExperiment={isDownloadsExperiment}
             isEditable={isEditable}
           />
         );
@@ -88,6 +90,7 @@ export type DownloadCardGroupProps = {
   control: Control<ResourceFormProps>;
   hasError?: boolean;
   triggerForm: () => void;
+  isDownloadsExperiment?: boolean;
 };
 const DownloadCardGroup: FC<DownloadCardGroupProps> = ({
   downloads,
@@ -95,6 +98,7 @@ const DownloadCardGroup: FC<DownloadCardGroupProps> = ({
   control,
   hasError = false,
   triggerForm,
+  isDownloadsExperiment = false,
 }) => {
   const groupedDownloads = groupDownloadResources(downloads, additionalFiles);
 
@@ -118,7 +122,10 @@ const DownloadCardGroup: FC<DownloadCardGroupProps> = ({
             <OakGrid
               $position="relative"
               $width="max-content"
-              $gridTemplateColumns={["1fr", "max-content max-content"]}
+              $gridTemplateColumns={[
+                "1fr",
+                `${isDownloadsExperiment ? "1fr 1fr" : "max-content max-content"}`,
+              ]}
               $cg={"space-between-s"}
               $rg={"space-between-s"}
             >
@@ -135,6 +142,7 @@ const DownloadCardGroup: FC<DownloadCardGroupProps> = ({
                     downloadType={downloadType}
                     triggerForm={triggerForm}
                     hasError={hasError}
+                    isDownloadsExperiment={isDownloadsExperiment}
                   />
                 );
               })}